TEST_SERVERS/PORTED_TO_SECUREMCP/memory-mcp/pyproject.toml
[project]
name = "memory-mcp"
version = "0.1.0"
description = "MongoDB Memory MCP server"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.36.26",
    "fastmcp>=2.6.1",
    "filetype>=1.2.0",
    "pillow>=11.2.1",
    "pydantic>=2.10.6",
    "pymongo>=4.12.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "tavily-python>=0.5.1",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = ["-v", "--tb=short", "--strict-markers"]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "performance: performance tests that may take longer",
    "integration: integration tests that test multiple components",
    "slow: tests that are known to be slow"
]
filterwarnings = [
    "ignore::pytest.PytestUnknownMarkWarning",
    "ignore::DeprecationWarning",
    "ignore::pytest_asyncio.plugin.PytestDeprecationWarning"
]